requirements.txt
pyramid
waitress
pillow
pytesseract
pdf2image